refactor: organize vsphere functions into smaller files
